+++ twitchio/user.py
@@ -784,6 +784,23 @@ class PartialUser:
 
         return Schedule(self._http, data)
 
+    async def fetch_channel_teams(self):
+        """|coro|
+
+        Fetches a list of Twitch Teams of which the specified channel/broadcaster is a member.
+
+        Returns
+        --------
+        List[:class:`twitchio.ChannelTeams`]
+        """
+        from .models import ChannelTeams
+
+        data = await self._http.get_channel_teams(
+            broadcaster_id=str(self.id),
+        )
+
+        return [ChannelTeams(self._http, x) for x in data]
+
 
 class BitLeaderboardUser(PartialUser):